+++ tests/testthat/test-collapsing.R
@@ -26,7 +26,7 @@ test_that("collapsing with formatting", {
 
 test_that("collapsing with formatting, custom seps", {
   expect_snapshot(local({
-    cli_div(theme = list(body = list(vec_sep = " ... ")))
+    cli_div(theme = list(div = list(vec_sep = " ... ")))
     pkgs <- paste0("pkg", 1:5)
     cli_text("Packages: {.pkg {pkgs}}.")
   }))